B7.1 Allow users to capture emails (text and attachments) as single records as well as individual records linked by metadata.
Solution :- Fetching stream of document associated with record on
creation of Electronic record first I am adding document which are
attachments of the email record and after storing document I am
declaring them as record and linking it with the email record.

B7.4 Ensure the capture of email transmission data as metadata
persistently linked to the email record. 

Solution : - We have to capture To and From field

B7.5 Ensure that the text of an email and its transmission details
cannot be amended in any way once the email has been captured. Nor
should the subject line of the email itself be changeable, although the
title of the record may be edited for easier access through, for
example, keywords or by file-naming conventions.

Solution:- Capturing Email body,subject and assigning it into the record

B7.6 Ensure that a human-readable version of an email message address is
also captured, where one exists. 

Already completed just we need to integrate with this project

Note:- We need to check once we will get the environment whether the attachments has been declared as record or not If not we need to declare